+++ Changelog.md
@@ -1,1 +1,3 @@
 - Fixed condensers not working
+- Added datamap override for neoforge fluid temperature
+- Added CraftTweaker integration (Thanks to DeusSixik)